requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration
ollama==0.1.6

# Command Execution & System
psutil==5.9.7
colorama==0.4.6

# Database
aiosqlite==0.19.0

# Logging & Monitoring
python-json-logger==2.0.7

# Security & Validation
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx>=0.25.2,<0.26.0
